Here are a large number of performance improvements. I want to move the discussion over from #47 as it's not possible to comment on actual code in an issue.

This code is not intended to be merged, while it is functional, it is primarily for discussion. I would rather wait until the code has been restructured into separate files, as that will make it easier to digest the changes.

Just to give an idea of the performance improvement [1](#1):

| Function | ops ↑ | ns/op ↓ | B/op ↓ | allocs/op ↓ |
| :---- | ----: | --: | --: | --: |
| **`Server.ReceivePacket()`** | |
| Original | 23,005 | 60,837 | 70,080 | 18 |
| New |  1,000,000 | 1,320 | 176 | 8 |
| **`ParsePacket()`** |  |
| Original | 121,566 | 8,235 | 4,624 | 18 |
| New |  1,000,000 | 1,802 | 304 | 10 |
| **`Message.String()`** | | | |
| Original | 804,878 | 3,510| 256 | 10 |
| New |  1,000,000 | 1,585 | 128 | 5 |
| **`Message.MarshalBinary()`** | |
| Original | 1,000,000 | 1,942 | 368 | 10 |
| New |  1,000,000 | 1,219 | 296 | 6 |
| New (Light Version) | 2,091,072 | 716 | 24 | 3 |

#####  [1]: https://replit.com/@chabad360/go-osc-benchmark