@chabad360 2021-10-08
I would love to, but I would kinda prefer to split `osc.go` into multiple separate files before finishing this up. It's a lot of changes, so it would be better to push this once the changes are more clear. However, if @hypebeast would prefer to do this first, I'll gladly finish it up.