@chabad360 2022-01-26
Oh, yes. I know about that, it bit me in back side pretty hard already. Very poor decision on my part (I don't actually remember why I made it do that). I used a temporary fix in my application, but in a week or two I'm going to be going on a bit of a bug hint/major rewrite. So it'll be fixed then.


I'll enable issue tracker later today.